reafactor: 코드 정리
