+++ src/pages/products/index.tsx
@@ -1,61 +1,29 @@
 import { useMemo, useState } from 'react'
 
-import { SORT_FIELD } from '@/constants/FILTERING_BROWSING'
 import useInfiniteCardsQuery from '@/hooks/api/productsAPI/useInfiniteCardsQuery'
+import useProductCardsParams from '@/hooks/api/productsAPI/useProductCardsParams'
 import useIntersectionObserver from '@/hooks/useIntersectionObserver'
 import AppBar from '@/pages/products/components/app-bar'
 import FilteringSheet from '@/pages/products/components/filtering-sheet'
 import SortOrdersSheet from '@/pages/products/components/sort-orders-sheet'
 import { SheetSliceState, sheet } from '@/store/sheet-slice.ts'
-import { ISearchProductsData } from '@/types/postProductData.type'
 
 import ProductCard from '@components/product-card'
 import ProductHeader from '@components/product-header'
 import { useDispatch, useSelector } from 'react-redux'
-import { useLocation } from 'react-router-dom'
 
 import * as S from './ProductsPage.style'
 
 const ProductsPage = () => {
   const dispatch = useDispatch()
-  const location = useLocation()
-  const queryParams = new URLSearchParams(location.search)
-  const input = queryParams.get('input')
-  const type = queryParams.get('type')
-  const minPrice = queryParams.get('minPrice')
-  const maxPrice = queryParams.get('maxPrice')
-  const startTime = queryParams.get('startTime')
-  const endTime = queryParams.get('endTime')
-  const date = queryParams.get('date')
-  const city = queryParams.get('city')
-  const sort = queryParams.get('sort')
-
-  const queryData: ISearchProductsData = useMemo(
-    () => ({
-      page: 0,
-      size: 6,
-      sortField: sort ? SORT_FIELD[sort] : undefined,
-      sortType: undefined,
-      keyword: input || undefined,
-      cityCode: city === '0' ? undefined : (city as string),
-      contentType:
-        type === '0' || type === undefined ? undefined : (type as string),
-      startDate: date || undefined,
-      endDate: date || undefined,
-      startTime: startTime || undefined,
-      endTime: endTime || undefined,
-      minPrice: Number(minPrice) || undefined,
-      maxPrice: Number(maxPrice) || undefined,
-    }),
-    [input, type, minPrice, maxPrice, date, city, sort],
-  )
+  const cardsQueryData = useProductCardsParams()
 
   const [isKebabClicked, setIsKebabClicked] = useState(false)
   const {
     data: cardData,
     hasNextPage,
     fetchNextPage,
-  } = useInfiniteCardsQuery(queryData)
+  } = useInfiniteCardsQuery(cardsQueryData)
 
   const cardsContents = useMemo(
     () => (cardData ? cardData.pages.flatMap((page) => page.content) : []),
